backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-JWT-Extended>=4.6.0
Werkzeug>=3.0.1
bcrypt>=4.1.2
Pillow>=10.2.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
